refactor: 패키지 infrastructure를 infra로 수정
패키지 이름이 너무 길어 줄임
