+++ src/main/java/org/example/cafe/infra/QuestionJdbcRepository.java
@@ -1,13 +1,13 @@
-package org.example.cafe.infrastructure;
+package org.example.cafe.infra;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 import org.example.cafe.domain.Question;
 import org.example.cafe.domain.QuestionRepository;
-import org.example.cafe.infrastructure.jdbc.GeneratedKeyHolder;
-import org.example.cafe.infrastructure.jdbc.JdbcTemplate;
-import org.example.cafe.infrastructure.jdbc.RowMapper;
+import org.example.cafe.infra.jdbc.GeneratedKeyHolder;
+import org.example.cafe.infra.jdbc.JdbcTemplate;
+import org.example.cafe.infra.jdbc.RowMapper;
 
 public class QuestionJdbcRepository implements QuestionRepository {
