+++ src/main/java/org/example/cafe/infra/jdbc/RowMapper.java
@@ -1,4 +1,4 @@
-package org.example.cafe.infrastructure.jdbc;
+package org.example.cafe.infra.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
