--- src/main/java/org/example/cafe/infrastructure/jdbc/exception/JdbcTemplateException.java
@@ -1,12 +1,0 @@
-package org.example.cafe.infrastructure.jdbc.exception;
-
-public class JdbcTemplateException extends RuntimeException {
-
-    public JdbcTemplateException(String message) {
-        super(message);
-    }
-
-    public JdbcTemplateException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
